o Classes I forgot to commit o Changed Money's amount from BigDecimal to Long.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/qi4j/libraries@8237 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ general/src/main/java/org/qi4j/library/general/model/Money.java
@@ -18,6 +18,6 @@ import java.math.BigDecimal;
  * Generic interface for Money which stores an amount and currency.
  * Both amount and currency must be immutable.
  */
-public interface Money extends Amount<BigDecimal>, Currency
+public interface Money extends Amount<Long>, Currency
 {
 }
